* fixed test

+++ APSEngine/src/test/java/elen7045/group5/project/aps/jpa/repository/NotificationRepositoryTest.java
@@ -43,7 +43,7 @@ public class NotificationRepositoryTest
 	public void testNotificationGetOne()
 	{
 
-		Integer testId = new Integer(3);
+		Integer testId = new Integer(2);
 		Notification notification = notificationService.findById(testId);
 
 		assertNotNull(notification);
